# api-schema/bin/main/openapi.yaml
openapi: 3.0.3
info:
  title: Connect Four API
  description: REST API for the Connect Four game
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: /api
    description: API base path

security: []

paths:
  /games:
    description: Endpoints for creating new Connect Four games
    post:
      summary: Create a new game
      operationId: createGame
      tags:
        - Games
      responses:
        "201":
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateGameResponse"
        "500":
          description: Server error

  /games/{id}:
    description: Endpoints for retrieving and deleting a specific game by ID
    get:
      summary: Get game state
      operationId: getGame
      tags:
        - Games
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Game ID
      responses:
        "200":
          description: Game state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameResponse"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete/abandon a game
      operationId: deleteGame
      tags:
        - Games
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Game ID
      responses:
        "204":
          description: Game deleted successfully
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /games/{id}/moves:
    description: Endpoints for making moves in a specific game
    post:
      summary: Make a move
      operationId: makeMove
      tags:
        - Games
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Game ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveRequest"
      responses:
        "200":
          description: Move executed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveResponse"
        "400":
          description: Invalid move
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    GameStatus:
      type: string
      enum:
        - ONGOING
        - PLAYER_WON
        - BOT_WON
        - DRAW
      description: Current state of the game

    BoardCellEnum:
      type: string
      enum:
        - "0"
        - "X"
        - "Y"
      description: Cell content (0=empty, X=player, Y=bot)

    Position:
      type: object
      required:
        - column
        - row
      properties:
        column:
          type: integer
          minimum: 0
          maximum: 6
          description: Column index (0-6)
        row:
          type: integer
          minimum: 0
          maximum: 5
          description: Row index (0-5)

    CreateGameResponse:
      type: object
      description: Response when creating a new game (just the ID since board is empty)
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          description: Unique game identifier

    GameResponse:
      type: object
      description: Full game state for reconnection
      required:
        - id
        - board
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique game identifier
        board:
          type: array
          items:
            type: array
            items:
              $ref: "#/components/schemas/BoardCellEnum"
          description: 6x7 game board
        status:
          $ref: "#/components/schemas/GameStatus"

    MoveRequest:
      type: object
      required:
        - column
      properties:
        column:
          type: integer
          minimum: 0
          maximum: 6
          description: Column to drop the piece in (0-6)

    MoveResponse:
      type: object
      description: Response after making a move (delta only)
      required:
        - status
      properties:
        botMove:
          $ref: "#/components/schemas/Position"
          description: Position where bot placed its piece (null if game ended before bot move)
        status:
          $ref: "#/components/schemas/GameStatus"

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - GAME_NOT_FOUND
            - INVALID_MOVE
            - GAME_OVER
            - SERVER_ERROR
          description: Error code
        message:
          type: string
          description: Human-readable error message